ReId-reid_implementation/C1/requirements.txt
# GPU PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO and computer vision
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Redis and utilities
redis>=4.6.0
numpy>=1.24.0